+++ src/routes/uploadresource.js
@@ -7,13 +7,21 @@ exports.route = "/uploadresource";
 
 exports.handler = function (req, res) {
 
+	var currentPath = '';
 	var form = new formidable.IncomingForm();
+
 	form.uploadDir = fs.realpathSync(path.join(process.cwd(), "data"));
 	form.keepExtensions = true;
 	form.parse(req);
 
+	form.on('field', function(name, value) {
+		if (name == 'path') {
+			currentPath = value;
+		}
+	});
+
 	form.on('file', function(name, file) {
-		var baseDir = fs.realpathSync(path.join(process.cwd(), "data", "resources"));
+		var baseDir = fs.realpathSync(path.join(process.cwd(), "data", "resources", currentPath));
 		var newfname = path.join(baseDir, file.name);
 		fs.rename(file.path, newfname , function (err) {
 			if ( err ) { console.log('ERROR: ' + err); }